Extra pedantic warnings fixed.
git-svn-id: https://nicola-bonelli-repo.googlecode.com/svn@855 0daf1152-ea34-11dd-b00f-9fabb4b2ff13

+++ trunk/codes.cpp11/matrix-test.cpp
@@ -289,7 +289,7 @@ Context(more_matrix_test)
 
     Test(det)
     {
-        more::matrix<double> a(3,3);
+        more::matrix<int> a(3,3);
         a = {1, 7, 3, 2, 5, 2, 3, 8, 10};
         Assert( det(a), is_equal_to(-61) );
     }
